refactor: 홈 화면 내부 슬라이더 컴포넌트 분리
